AI Customer Support Bot/pyproject.toml
[project]
name = "ai-customer-support-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.6.9",
    "ddgs>=9.5.4",
    "numpy>=2.2.6",
    "openai-agents>=0.2.8",
    "opencv-python>=4.12.0.88",
    "pdfplumber>=0.11.7",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "tesseract>=0.1.3",
]
